Refactor of VirtualNode-controller.
This commit allows to manage in a simpler way the NamespaceMap lifecycle.
